+++ src/components/Main/Company/Table/RowDetail/DetailView/DetailView.styles.jsx
@@ -1,6 +1,6 @@
 import { styled } from "styled-components";
-import { colorHandler } from "../Table.utils.js";
-import { appgrey, apporange } from "../../../../../utils/colors.styles.jsx";
+import { colorHandler } from "../../Table.utils";
+import { appgrey, apporange } from "../../../../../../utils/colors.styles";
 
 export const $RowForm = styled.form`
     display: block;
@@ -11,12 +11,16 @@ export const $RowDetailsHeader = styled.div`
     justify-content: space-between;
 `;
 
-export const $RowDetailsBox = styled.div`
+export const $RowInputRange = styled.input`
+    cursor: pointer;
+`
+
+export const $RowStatusLabel = styled.div`
     display: flex;
-    justify-content: ${({$justify}) => $justify};
+    justify-content: space-evenly;
     align-items: center;
-    gap: ${({$gap}) => $gap};
-    width: 100%;
+    color: ${({$color}) => colorHandler($color)};
+    padding: 10px 0 0 0;
 `
 
 export const $RowInputField = styled.input`
@@ -31,6 +35,14 @@ export const $RowInputField = styled.input`
       outline: none;
       border: 1px solid ${apporange};
     }
+`
+
+export const $RowDetailsBox = styled.div`
+    display: flex;
+    justify-content: ${({$justify}) => $justify};
+    align-items: center;
+    gap: ${({$gap}) => $gap};
+    width: 100%;
 `
 
 export const $RowButton = styled.button`
@@ -51,18 +63,4 @@ export const $RowButton = styled.button`
     & :nth-child(2) {
         font-size: 12px;
     }
-`
-
-export const $RowStatusLabel = styled.div`
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    color: ${({$color}) => colorHandler($color)};
-    padding: 10px 0 0 0;
-`
-
-export const $Childs = styled.div`
-    display: flex;
-    align-items: center;
-    padding: 10px 0;
 `